backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.27
aiosqlite==0.20.0
yt-dlp==2024.3.10
spacy==3.7.4
nltk==3.8.1
slowapi==0.1.9
pydantic==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
aiofiles==23.2.1
